requirements.txt
blinker ==            1.5
certifi ==            2022.12.7
cffi ==               1.15.1
charset-normalizer == 3.1.0
click ==              8.1.3
colorama ==           0.4.5
cryptography ==       39.0.2
Flask ==             2.2.3
Flask-JWT-Extended == 4.4.4
Flask-Login ==       0.6.2
Flask-Mail  ==       0.9.1
Flask-MySQL ==        1.5.2
Flask-MySQLdb ==      1.0.1
Flask-WTF ==          1.0.1
greenlet  ==          1.1.3
idna ==               3.4
itsdangerous ==       2.1.2
Jinja2 ==             3.1.2
MarkupSafe ==         2.1.1
mysql-connector ==    2.2.9
mysqlclient ==        2.1.1
pip ==                23.0.1
pycparser ==          2.21
PyJWT ==              2.6.0
PyMySQL ==            1.0.2
python-dotenv ==      1.0.0
setuptools ==         63.2.0
urllib3 ==            1.26.15
Werkzeug ==           2.2.2